Pushing to Bridgitte

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountsDao.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public interface AccountsDao {
 
 //return balance
-    Integer getBalance(String username);
+    BigDecimal getBalance(String username);
 
 
 }
